# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация к API.",
        "title": "Gangchi API",
        "contact": {},
        "version": "0.2"
    },
    "basePath": "/",
    "paths": {
        "/api/blog/categories": {
            "get": {
                "description": "Возвращает все категории блога с переводом",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Блог"
                ],
                "summary": "Получить список категорий",
                "operationId": "get_blog_categories",
                "parameters": [
                    {
                        "type": "string",
                        "default": "ru",
                        "description": "Язык: ru, en, kk, zh",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BlogCategory"
                            }
                        }
                    }
                }
            }
        },
        "/api/blog/posts": {
            "get": {
                "description": "Возвращает список статей с пагинацией, фильтрацией по категориям и переводом",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Блог"
                ],
                "summary": "Получить список статей",
                "operationId": "get_blog_posts",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Страница",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "На странице",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Слаги категорий (можно несколько)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "ru",
                        "description": "Язык: ru, en, kk, zh",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BlogPostsResponse"
                        }
                    }
                }
            }
        },
        "/api/blog/posts/{slug}": {
            "get": {
                "description": "Возвращает полную статью с тегами, переводом и SEO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Блог"
                ],
                "summary": "Получить статью по слагy",
                "operationId": "get_blog_post_by_slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Слаг статьи",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "ru",
                        "description": "Язык: ru, en, kk, zh",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BlogPostResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/loaders/getall": {
            "get": {
                "description": "Отправляет всю таблицу с данными о каждом погрузчике",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Электрические погрузчики"
                ],
                "summary": "Получить все погрузчики",
                "operationId": "all_loaders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Языковой код",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Loader"
                            }
                        }
                    }
                }
            }
        },
        "/api/loaders/getimage/{loader_id}": {
            "get": {
                "description": "Отправляет фото погрузчика с указанным id",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "Электрические погрузчики"
                ],
                "summary": "Получить фото погрузчика по id",
                "operationId": "loader_image",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1",
                        "description": "ID погрузчика",
                        "name": "loader_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/manual_loaders/getall": {
            "get": {
                "description": "Отправляет всю таблицу с данными о каждом погрузчике",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ручные погрузчики"
                ],
                "summary": "Получить все ручные погрузчики",
                "operationId": "all_manual_loaders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Языковой код",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ManualLoader"
                            }
                        }
                    }
                }
            }
        },
        "/api/manual_loaders/getimage/{loader_id}": {
            "get": {
                "description": "Отправляет фото погрузчика с указанным id",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "Ручные погрузчики"
                ],
                "summary": "Получить фото ручного погрузчика по id",
                "operationId": "get_manual_loader_image",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1",
                        "description": "ID ручного погрузчика",
                        "name": "loader_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BlogAuthor": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.BlogCategory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "заполняется переводом",
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "models.BlogPostDetail": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.BlogAuthor"
                },
                "category": {
                    "$ref": "#/definitions/models.BlogCategory"
                },
                "content": {
                    "type": "string"
                },
                "coverImage": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "publishedAt": {
                    "type": "string"
                },
                "seo": {
                    "$ref": "#/definitions/models.BlogSEO"
                },
                "slug": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BlogTag"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.BlogPostListItem": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.BlogAuthor"
                },
                "category": {
                    "$ref": "#/definitions/models.BlogCategory"
                },
                "coverImage": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "publishedAt": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.BlogPostResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.BlogPostDetail"
                }
            }
        },
        "models.BlogPostsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BlogPostListItem"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/models.PaginationMeta"
                }
            }
        },
        "models.BlogSEO": {
            "type": "object",
            "properties": {
                "metaDescription": {
                    "type": "string"
                },
                "metaTitle": {
                    "type": "string"
                }
            }
        },
        "models.BlogTag": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "models.Loader": {
            "type": "object",
            "properties": {
                "autoWeight": {
                    "type": "integer"
                },
                "batteryType": {
                    "type": "string"
                },
                "brakeType": {
                    "type": "string"
                },
                "chargingTime": {
                    "type": "string"
                },
                "engineType": {
                    "type": "string"
                },
                "forkLength": {
                    "type": "integer"
                },
                "frontWheels": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "hydraulicLiftingEngine": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "length": {
                    "type": "integer"
                },
                "liftHeight": {
                    "type": "integer"
                },
                "liftingAngle": {
                    "type": "integer"
                },
                "liftingCylinder": {
                    "type": "string"
                },
                "longmenFrameMaterial": {
                    "type": "integer"
                },
                "maxLiftWeight": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "picturePathLinux": {
                    "type": "string"
                },
                "picturePathWindows": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "rearWheels": {
                    "type": "string"
                },
                "steeringMode": {
                    "type": "string"
                },
                "turningRadius": {
                    "type": "integer"
                },
                "voltage": {
                    "type": "integer"
                },
                "wheelAxis": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                },
                "workingHours": {
                    "type": "string"
                }
            }
        },
        "models.ManualLoader": {
            "type": "object",
            "properties": {
                "brakeType": {
                    "type": "string"
                },
                "control": {
                    "type": "string"
                },
                "driveGear": {
                    "type": "string"
                },
                "forkLength": {
                    "type": "integer"
                },
                "forkWidth": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "length": {
                    "type": "integer"
                },
                "liftingSpeed": {
                    "type": "integer"
                },
                "maxLiftWeight": {
                    "type": "integer"
                },
                "maxSpeed": {
                    "type": "number",
                    "format": "float64"
                },
                "name": {
                    "type": "string"
                },
                "picturePathLinux": {
                    "type": "string"
                },
                "picturePathWindows": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "models.PaginationMeta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "perPage": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        }
    }
}